black-list-app/src/main/java/lv/javaguru/black/list/core/validations/ValidatePersonFirstNameNotNullOrBlank.java
package lv.javaguru.black.list.core.validations;

import java.util.Optional;
import lombok.RequiredArgsConstructor;
import lv.javaguru.black.list.core.api.dto.PersonDTO;
import lv.javaguru.black.list.core.api.dto.ValidationErrorDTO;
import org.springframework.stereotype.Component;

@Component
@RequiredArgsConstructor(access = lombok.AccessLevel.PACKAGE)
class ValidatePersonFirstNameNotNullOrBlank implements ValidatePersonField {

    private final ValidationErrorFactory validationErrorFactory;

    @Override
    public Optional<ValidationErrorDTO> validate(PersonDTO person) {
        return (person.personFirstName() == null || person.personFirstName().isBlank())
                ? Optional.of(validationErrorFactory.buildError("ERROR_CODE_1"))
                : Optional.empty();
    }

}
